server/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
protobuf==5.29.1
psycopg2==2.9.10
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
